requirements.txt
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
curl_cffi==0.13.0
distro==1.9.0
eval_type_backport==0.2.2
filelock==3.20.0
Flask==3.1.2
frozendict==2.4.6
fsspec==2025.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.11
invoke==2.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.0
langchain-core==1.0.0
langgraph==1.0.0
langgraph-checkpoint==2.1.2
langgraph-prebuilt==1.0.0
langgraph-sdk==0.2.9
langsmith==0.4.37
Markdown==3.9
MarkupSafe==3.0.3
mistralai==0.4.2
mpmath==1.3.0
multitasking==0.0.12
networkx==3.5
numpy==2.3.4
openai==2.5.0
orjson==3.10.18
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
peewee==3.18.2
platformdirs==4.5.0
protobuf==6.33.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.6.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.6.0
yfinance==0.2.66
zstandard==0.25.0
gunicorn